Event: Move event aliases to deprecated
Fixes gh-3214
